restructured solution a little bit
