+++ src/Hive.IO/Hive.IO/Forms/ConversionTechPropertiesViewModel.cs
@@ -6,7 +6,7 @@ using System.Windows;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using Hive.IO.EnergySystems;
-using Hive.IO.Resources;
+using Hive.IO.Util;
 using OxyPlot.Reporting;
 using Rhino.Geometry;
