+++ src/Hive.IO/Hive.IO/Util/Misc.cs
